LPS-95007 Move OrganizationImplTest of the com.liferay.portal.model.impl package to organizations-test module
